pedestrians routeのerror変換を分離
